Setting subject for contact

+++ controller/PostsControl.php
@@ -33,148 +33,149 @@ class PostsControl
 
     public function createChapter()
     {
+        if (isset($_SESSION['isAdmin'])) {
+            if ($_SESSION['isAdmin'] = true) {
 
-        $_SESSION['author_post_title'] = "";
-        $_SESSION['author_post_content'] = "";
 
-        
-        if (isset($_POST['author_post_title'], $_POST['author_post_content'])) 
-        {
-            
+                $_SESSION['author_post_title'] = "";
+                $_SESSION['author_post_content'] = "";
 
-            if (!empty($_POST['author_post_title']) and !empty($_POST['author_post_content'])) 
-            {
-                $titleChapter = $_POST['author_post_title'];
-                $_SESSION['author_post_title'] = $titleChapter;
 
-                $contentChapter = $_POST['author_post_content'];
-                $_SESSION['author_post_content'] = $contentChapter;
+                if (isset($_POST['author_post_title'], $_POST['author_post_content'])) {
 
-                if (isset($_FILES['image_chapter']) and !empty($_FILES['image_chapter']['name'])) 
-                {
 
-                    // Define constants to verify uploaded file
-                    $maxWeightFile = 2097152;
-                    $validExtensions = array('jpg', 'jpeg', 'png', 'gif');
-                    $imageInfos = pathinfo($_FILES['image_chapter']['name']);
-                    $uploadedExtension = $imageInfos['extension'];
-                    $imageFile = '' . time() . '.' . $uploadedExtension;
+                    if (!empty($_POST['author_post_title']) and !empty($_POST['author_post_content'])) {
+                        $titleChapter = $_POST['author_post_title'];
+                        $_SESSION['author_post_title'] = $titleChapter;
 
-                    $pathToUpload = ROOT . 'public/images/chapters/';
+                        $contentChapter = $_POST['author_post_content'];
+                        $_SESSION['author_post_content'] = $contentChapter;
 
-                    if ($_FILES['image_chapter']['size'] <= $maxWeightFile) 
-                    {
-                        if (in_array($uploadedExtension, $validExtensions)) 
-                        {
-                            $imageChapter = move_uploaded_file($_FILES['image_chapter']['tmp_name'], $pathToUpload . $imageFile);
+                        if (isset($_FILES['image_chapter']) and !empty($_FILES['image_chapter']['name'])) {
 
-                            if ($imageChapter) 
-                            {
-                                $newImageFile = $imageFile;
-                                $createContent = new \JForteroche\Blog\Model\PostManager();
-                                $newEntry = $createContent->newPost($titleChapter, $contentChapter, $newImageFile);
+                            // Define constants to verify uploaded file
+                            $maxWeightFile = 2097152;
+                            $validExtensions = array('jpg', 'jpeg', 'png', 'gif');
+                            $imageInfos = pathinfo($_FILES['image_chapter']['name']);
+                            $uploadedExtension = $imageInfos['extension'];
+                            $imageFile = '' . time() . '.' . $uploadedExtension;
 
-                                $_SESSION['success'] = 1;
-                                $_SESSION['success_upload'] = 'Votre chapitre a été publié avec succès !';
-                                $_SESSION['author_post_title'] = "";
-                                $_SESSION['author_post_content'] = "";
-                                header('Location: ' . HOST . 'admin/dashboard');
+                            $pathToUpload = ROOT . 'public/images/chapters/';
 
+                            if ($_FILES['image_chapter']['size'] <= $maxWeightFile) {
+                                if (in_array($uploadedExtension, $validExtensions)) {
+                                    $imageChapter = move_uploaded_file($_FILES['image_chapter']['tmp_name'], $pathToUpload . $imageFile);
+
+                                    if ($imageChapter) {
+                                        $newImageFile = $imageFile;
+                                        $createContent = new \JForteroche\Blog\Model\PostManager();
+                                        $newEntry = $createContent->newPost($titleChapter, $contentChapter, $newImageFile);
+
+                                        $_SESSION['success'] = 1;
+                                        $_SESSION['success_upload'] = 'Votre chapitre a été publié avec succès !';
+                                        $_SESSION['author_post_title'] = "";
+                                        $_SESSION['author_post_content'] = "";
+                                        header('Location: ' . HOST . 'admin/dashboard');
+                                    }
+                                } else {
+                                    $_SESSION['success'] = 0;
+                                    header('Location: ' . HOST . 'admin/create');
+                                    $_SESSION['error_upload'] = 'Votre photo doit être au format jpg, jpeg ou png.';
+                                    exit();
+                                }
+                            } else {
+                                $_SESSION['success'] = 0;
+                                header('Location: ' . HOST . 'admin/create');
+                                $_SESSION['error_upload'] = 'Votre photo ne doit pas dépaser 2Mo.';
+                                exit();
                             }
                         } else {
                             $_SESSION['success'] = 0;
                             header('Location: ' . HOST . 'admin/create');
-                            $_SESSION['error_upload'] = 'Votre photo doit être au format jpg, jpeg ou png.';
+                            $_SESSION['error_upload'] = 'Vous devez sélectionner une image à joindre à votre chapitre.';
                             exit();
                         }
                     } else {
                         $_SESSION['success'] = 0;
                         header('Location: ' . HOST . 'admin/create');
-                        $_SESSION['error_upload'] = 'Votre photo ne doit pas dépaser 2Mo.';
+                        $_SESSION['error_upload'] = 'Vous devez donner un titre, un contenu et une image à votre chapitre.';
                         exit();
                     }
-                } else {
-                    $_SESSION['success'] = 0;
-                    header('Location: ' . HOST . 'admin/create');
-                    $_SESSION['error_upload'] = 'Vous devez sélectionner une image à joindre à votre chapitre.';
-                    exit();
                 }
             } else {
-                $_SESSION['success'] = 0;
-                header('Location: ' . HOST . 'admin/create');
-                $_SESSION['error_upload'] = 'Vous devez donner un titre, un contenu et une image à votre chapitre.';
-                exit();
+                header('Location: ' . HOST . 'connexion');
             }
         }
-        
     }
 
 
 
     public function updateChapter()
     {
-        if (isset($_POST['author_post_title'], $_POST['author_post_content'])) 
-        {
-            if (!empty($_POST['author_post_title']) and !empty($_POST['author_post_content'])) 
-            {
-                $titleChapter = $_POST['author_post_title'];
-                $_SESSION['author_post_title'] = $titleChapter;
+        if (isset($_SESSION['isAdmin'])) {
+            if ($_SESSION['isAdmin'] = true) {
 
-                $contentChapter = $_POST['author_post_content'];
-                $_SESSION['author_post_content'] = $contentChapter;
+                if (isset($_POST['submit_edited_chapter'])) {
+                    if (isset($_POST['author_post_title'], $_POST['author_post_content'])) {
+                        if (!empty($_POST['author_post_title']) and !empty($_POST['author_post_content'])) {
+                            $titleChapter = $_POST['author_post_title'];
 
-                if (isset($_FILES['image_chapter']) and !empty($_FILES['image_chapter']['name'])) 
-                {
+                            $contentChapter = $_POST['author_post_content'];
 
-                    // Define constants to verify uploaded file
-                    $maxWeightFile = 2097152;
-                    $validExtensions = array('jpg', 'jpeg', 'png', 'gif');
-                    $imageInfos = pathinfo($_FILES['image_chapter']['name']);
-                    $uploadedExtension = $imageInfos['extension'];
-                    $imageFile = '' . time() . '.' . $uploadedExtension;
+                            if (isset($_FILES['image_chapter']) and !empty($_FILES['image_chapter']['name'])) {
 
-                    $pathToUpload = ROOT . 'public/images/chapters/';
+                                // Define constants to verify uploaded file
+                                $maxWeightFile = 2097152;
+                                $validExtensions = array('jpg', 'jpeg', 'png', 'gif');
+                                $imageInfos = pathinfo($_FILES['image_chapter']['name']);
+                                $uploadedExtension = $imageInfos['extension'];
+                                $imageFile = '' . time() . '.' . $uploadedExtension;
 
-                    if ($_FILES['image_chapter']['size'] <= $maxWeightFile) 
-                    {
-                        if (in_array($uploadedExtension, $validExtensions)) 
-                        {
-                            $imageChapter = move_uploaded_file($_FILES['image_chapter']['tmp_name'], $pathToUpload . $imageFile);
+                                $pathToUpload = ROOT . 'public/images/chapters/';
 
-                            if ($imageChapter) 
-                            {
-                                $newImageFile = $imageFile;
-                                $insertContent = new \JForteroche\Blog\Model\PostManager();
-                                $insertContent->updatePost($_POST['postId'], $titleChapter, $contentChapter, $newImageFile);
+                                if ($_FILES['image_chapter']['size'] <= $maxWeightFile) {
+                                    if (in_array($uploadedExtension, $validExtensions)) {
+                                        $imageChapter = move_uploaded_file($_FILES['image_chapter']['tmp_name'], $pathToUpload . $imageFile);
 
-                                $_SESSION['success'] = 1;
-                                $_SESSION['success_upload'] = 'Votre chapitre a été modifié avec succès !';
-                                header('Location:' . HOST . 'admin/dashboard');
+                                        if ($imageChapter) {
+                                            $newImageFile = $imageFile;
+                                            $insertContent = new \JForteroche\Blog\Model\PostManager();
+                                            $insertContent->updatePost($_POST['postId'], $titleChapter, $contentChapter, $newImageFile);
+
+                                            $_SESSION['success'] = 1;
+                                            $_SESSION['success_upload'] = 'Votre chapitre a été modifié avec succès !';
+                                            $_SESSION['author_post_title'] = "";
+                                            $_SESSION['author_post_content'] = "";
+                                            header('Location:' . HOST . 'admin/dashboard');
+                                        }
+                                    } else {
+                                        $_SESSION['success'] = 0;
+                                        header('Location: ' . HOST . 'admin/edit-post&id=' . $_SESSION['id_chapter']);
+                                        $_SESSION['error_upload'] = 'Votre photo doit être au format jpg, jpeg ou png.';
+                                        exit();
+                                    }
+                                } else {
+                                    $_SESSION['success'] = 0;
+                                    header('Location: ' . HOST . 'admin/edit-post&id=' . $_SESSION['id_chapter']);
+                                    $_SESSION['error_upload'] = 'Votre photo ne doit pas dépaser 2Mo.';
+                                    exit();
+                                }
+                            } else {
+                                $_SESSION['success'] = 0;
+                                header('Location: ' . HOST . 'admin/edit-post&id=' . $_SESSION['id_chapter']);
+                                $_SESSION['error_upload'] = 'Vous devez sélectionner une image à joindre à votre chapitre.';
+                                exit();
                             }
-
                         } else {
                             $_SESSION['success'] = 0;
-                            header('Location: ' . $_SERVER['HTTP_REFERER']);
-                            $_SESSION['error_upload'] = 'Votre photo doit être au format jpg, jpeg ou png.';
+                            header('Location: ' . HOST . 'admin/edit-post&id=' . $_SESSION['id_chapter']);
+                            $_SESSION['error_upload'] = 'Vous devez donner un titre, un contenu et une image à votre chapitre.';
                             exit();
                         }
-                    } else {
-                        $_SESSION['success'] = 0;
-                        header('Location: ' . $_SERVER['HTTP_REFERER']);
-                        $_SESSION['error_upload'] = 'Votre photo ne doit pas dépaser 2Mo.';
-                        exit();
                     }
-                } else {
-                    $_SESSION['success'] = 0;
-                    header('Location: ' . $_SERVER['HTTP_REFERER']);
-                    $_SESSION['error_upload'] = 'Vous devez sélectionner une image à joindre à votre chapitre.';
-                    exit();
                 }
             } else {
-                $_SESSION['success'] = 0;
-                header('Location: ' . $_SERVER['HTTP_REFERER']);
-                $_SESSION['error_upload'] = 'Vous devez donner un titre, un contenu et une image à votre chapitre.';
-                exit();
+                header('Location: ' . HOST . 'connexion');
             }
         }
     }
@@ -183,25 +184,39 @@ class PostsControl
 
     public function editChapter()
     {
-        $postManager = new \JForteroche\Blog\Model\PostManager();
-        $post = $postManager->getPost($_GET['id']);
-        $_SESSION['id_chapter'] = $_GET['id'];
+        if (isset($_SESSION['isAdmin'])) {
+            if ($_SESSION['isAdmin'] = true) {
+
+                $postManager = new \JForteroche\Blog\Model\PostManager();
+                $post = $postManager->getPost($_GET['id']);
+                $_SESSION['id_chapter'] = $_GET['id'];
 
 
-        $viewToDisplay = new ViewRenderer('editView');
-        $viewToDisplay->renderView(array('post' => $post));
+                $viewToDisplay = new ViewRenderer('editView');
+                $viewToDisplay->renderView(array('post' => $post));
+            } else {
+                header('Location: ' . HOST . 'connexion');
+            }
+        }
     }
 
 
 
     public function deleteChapter()
     {
-        $id = htmlspecialchars($_GET['id']);
+        if (isset($_SESSION['isAdmin'])) {
+            if ($_SESSION['isAdmin'] = true) {
 
-        $postManager = new \JForteroche\Blog\Model\PostManager();
-        // Vérifier que l'action de suppression avec un message d'alerte avant de valider le traitement de la suppression du post de la db
-        $post = $postManager->deletePost($id);
+                $id = htmlspecialchars($_GET['id']);
 
-        header('Location:' . HOST . 'admin/dashboard');
+                $postManager = new \JForteroche\Blog\Model\PostManager();
+                // Vérifier que l'action de suppression avec un message d'alerte avant de valider le traitement de la suppression du post de la db
+                $post = $postManager->deletePost($id);
+
+                header('Location:' . HOST . 'admin/dashboard');
+            }
+        } else {
+            header('Location: ' . HOST . 'connexion');
+        }
     }
 }
